Corrects imports excluding jcommander [WIP] Refactor QRCodeDataGenerator
